services/websocket-service/pyproject.toml
[project]
name = "websocket-service"
version = "0.1.0"
description = "WebSocket Service for real-time task synchronization"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "dapr>=1.14.0",
    "dapr-ext-fastapi>=1.14.0",
    "websockets>=13.0",
    "httpx>=0.27.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.3.0",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
